fix(d3d11va): width-safe bitstream buffer bound check (HIGH memory-safety)
copy_into_buffer guarded the memcpy into the GPU-mapped decoder buffer with
(data.len() as u32) > buf_size. The as u32 cast TRUNCATES on a 64-bit host: a
>4 GiB bitstream whose low 32 bits land below buf_size passed the check, then
copy_nonoverlapping(.., data.len()) copied the full length into a buf_size-byte
GPU buffer -> OOB write / UB. The bitstream length is attacker-influenced
(file/item size). Compare in usize (buf_size as usize) before the cast, matching
the safe u32::try_from(..).unwrap_or(MAX) pattern used elsewhere in the file.

Refs CLAUDE.md 2026-05-31 review. Cross-compiles clean for x86_64-pc-windows-gnu.

+++ heic-backend-d3d11va/src/decoder.rs
@@ -505,7 +505,11 @@ fn copy_into_buffer(
     // the requested buffer slot.
     unsafe { ctx.GetDecoderBuffer(decoder, buffer_type, &mut buf_size, &mut buf_ptr) }
         .map_err(|e| BackendError::Decode(format!("GetDecoderBuffer({buffer_type:?}): {e}")))?;
-    if buf_ptr.is_null() || (data.len() as u32) > buf_size {
+    // Compare in usize: `data.len() as u32` TRUNCATES on a 64-bit host, so a
+    // >4 GiB bitstream whose low 32 bits land below buf_size would pass this
+    // guard and the copy_nonoverlapping below would overflow the GPU buffer
+    // (OOB write / UB). `buf_size as usize` widens losslessly.
+    if buf_ptr.is_null() || data.len() > buf_size as usize {
         // SAFETY: pairs with GetDecoderBuffer; harmless on null.
         unsafe { ctx.ReleaseDecoderBuffer(decoder, buffer_type) }
             .map_err(|e| BackendError::Decode(format!("ReleaseDecoderBuffer: {e}")))?;
